test(miot): 对齐层测试挪出 miot/ 子目录，否则 CI 收集不到
backend/pyproject.toml 的 norecursedirs 里 "miot" 不含斜杠，pytest 对这种
pattern 按目录名匹配任意深度，新建的 miloco/tests/miot/ 就跟着 backend/miot/
那个 workspace 包一起被跳过了。CI 跑的是不带路径参数的全量 pytest，这批测试
从未执行过；本地按路径跑不受收集规则约束，所以一直是绿的。

不动那条规则：它是为 backend/miot/ 写的，改了会连带改变本 PR 范围之外的收集
行为，也会触发依赖变更的放行流程。改成挪文件，跟根下已有的 test_miot_*.py 同
一个惯例。

result_codes 的三条并进已有的 test_result_codes.py，同一个模块不再分两处测。
